Reorganizing desing_new_data and some small changes
