+++ examples/spinodal2D.jl
@@ -2,12 +2,12 @@
 using Jecco.AdS5_3_1
 
 grid = SpecCartGrid3D(
-    x_min            = -20.,
-    x_max            =  20.,
-    x_nodes          =  150,
-    y_min            = -20.,
-    y_max            =  20.,
-    y_nodes          =  150,
+    x_min            = -50.,
+    x_max            =  50.,
+    x_nodes          =  125,
+    y_min            = -50.,
+    y_max            =  50.,
+    y_nodes          =  125,
     u_outer_min      =  0.1,
     u_outer_max      =  1.005,
     u_outer_domains  =  1,
@@ -16,6 +16,7 @@ grid = SpecCartGrid3D(
     fd_order         =  4,
     sigma_diss       =  0.2,
 )
+
 
 phiM2 = -(0.85)^2
 phiQ  = 10.
@@ -29,7 +30,7 @@ potential = AdS5_3_1.Phi8Potential(
 )
 
 id = BlackBranePert(
-    energy_dens = 0.133,
+    energy_dens = .209,
     phi0        = 1.0,
     phi2        = 0.9,
     oophiM2     = potential.oophiM2,
@@ -39,7 +40,7 @@ id = BlackBranePert(
     #a4_ampy     = -0.05,
     #a4_ky       = 1,
     xi0         = 0.02,
-    AH_pos      = 0.95,
+    AH_pos      = 0.9,
     xmax        = grid.x_max,
     xmin        = grid.x_min,
     ymin        = grid.y_min,
@@ -53,15 +54,15 @@ evoleq = AffineNull(
 )
 
 diag = DiagAH(
-    find_AH_every_t    = 0.5,
+    find_AH_every_t    = 0.1,
 )
 
 outdir = "/home/mikel/Documents/Jecco.jl/data/test/"
 
 io = InOut(
-    out_boundary_every_t        = .5,
-    out_bulk_every_t            = .5,
-    out_gauge_every_t           = .5,
+    out_boundary_every_t        = 1.,
+    out_bulk_every_t            = 1.,
+    out_gauge_every_t           = 1.,
     #out_bulkconstrained_every_t = 5.0,
     checkpoint_every_walltime_hours = 1,
     out_dir                     = outdir,
