+++ scripts/create_new_initial.jl
@@ -3,19 +3,19 @@ using FFTW
 using Plots
 gr()
 
-dirname   = "/home/mikel/Documents/Jecco.jl/data/bubbles/phiM_0.85_phiQ_10/bubble/"
+dirname   = "/home/mikel/Documents/Jecco.jl/data/end_data/"
 outdir    = "/home/mikel/Documents/Jecco.jl/data/new_data/"
 A_dir     = "/home/mikel/Documents/Jecco.jl/data/bubbles/phiM_0.85_phiQ_10/state_A_e_1.318/"
 B_dir     = "/home/mikel/Documents/Jecco.jl/data/bubbles/phiM_0.85_phiQ_10/state_B_e_0.133/"
-PS_dir    = "/home/mikel/Documents/Jecco.jl/data/bubbles/phiM_0.85_phiQ_10/phase_separated/e_1.0_L_40/"
+PS_dir    = "/home/mikel/Documents/Jecco.jl/data/bubbles/phiM_0.85_phiQ_10/phase_separated/e_1.0_AH_0.95/"
 
 grid = SpecCartGrid3D(
-    x_min            = -20.,
-    x_max            =  20.,
-    x_nodes          =  150,
-    y_min            = -20.,
-    y_max            =  20.,
-    y_nodes          =  150,
+    x_min            = -50.,
+    x_max            =  50.,
+    x_nodes          =  125,
+    y_min            = -50.,
+    y_max            =  50.,
+    y_nodes          =  125,
     u_outer_min      =  0.1,
     u_outer_max      =  1.005,
     u_outer_domains  =  1,
@@ -39,8 +39,8 @@ potential = AdS5_3_1.Phi8Potential(
 io = InOut(recover_dir = dirname, out_dir = outdir, checkpoint_dir = outdir,
            out_boundary_every=1, out_gauge_every=1,out_bulk_every=1,remove_existing = true,)
 
-new_center = (10.,10.)
-e_new      = 1.82
+new_center = (10., 10.)
+e_new      = 1.7
 
 #=
 parameters = AdS5_3_1.new_parameters(
@@ -72,16 +72,16 @@ parameters_collision =AdS5_3_1.new_parameters_coll(
 #AdS5_3_1.create_checkpoint(io, potential)
 #AdS5_3_1.initial_numerical_phi(grid, io, potential)
 #AdS5_3_1.shift(io, potential, new_center=new_center)
-#AdS5_3_1.new_box(grid, io, potential, same_spacing=:no)
-#AdS5_3_1.change_energy(io, e_new, potential, fix=:high)
+AdS5_3_1.new_box(grid, io, potential, same_spacing=:no)
+#AdS5_3_1.change_energy(io, e_new, potential, fix=:no)
 #AdS5_3_1.to1plus1(grid, io, potential)
 #AdS5_3_1.to2plus1(io, potential)
-#AdS5_3_1.create_new_data(grid, io, parameters, potential)
 #AdS5_3_1.design_collision(grid, io, parameters_collision)
 #AdS5_3_1.bubble_expansion(grid, io, potential, A_dir, B_dir, PS_dir,
-#                                    same_spacing=:no, b_cold=false)
+#                                  same_spacing=:no, b_cold=false)
 #AdS5_3_1.join_boxes(io, potential, dirname, dirname)
-AdS5_3_1.cut_circular_hole(io, potential, 6.0, 100, 100)
+#AdS5_3_1.cut_1D_hole(io, potential, 3.0, 50)
+#AdS5_3_1.cut_circular_hole(io, potential, 1.0, 70, 70)
 
 #=
 phi11 = BulkTimeSeries(dirname,:phi,1)
@@ -95,23 +95,23 @@ convert_to_mathematica(io.out_dir)
 
 e     = VEVTimeSeries(outdir, :energy)
 px    = VEVTimeSeries(outdir, :px)
-e_old = VEVTimeSeries(dirname, :energy)
+#e_old = VEVTimeSeries(dirname, :energy)
 
-#e_A   = VEVTimeSeries(A_dir, :energy)
-#e_B   = VEVTimeSeries(B_dir, :energy)
-#e_PS  = VEVTimeSeries(PS_dir, :energy)
-Jx = VEVTimeSeries(outdir,:Jx)
-Jy = VEVTimeSeries(outdir,:Jy)
+e_A   = VEVTimeSeries(A_dir, :energy)
+e_B   = VEVTimeSeries(B_dir, :energy)
+e_PS  = VEVTimeSeries(PS_dir, :energy)
+#Jx = VEVTimeSeries(outdir,:Jx)
+#Jy = VEVTimeSeries(outdir,:Jy)
 
 t,x,y  = get_coords(e,:,:,:)
-t_old, x_old, y_old = get_coords(e_old,:,:,:)
+#t_old, x_old, y_old = get_coords(e_old,:,:,:)
 plan   = plan_rfft(e[1,:,:])
 Nx     = length(x)
 Ny     = length(y)
 e0     = real(1/(Nx*Ny) * (plan * e[1,:,:]))[1]
 #e0_old = real(1/(Nx*Ny) * (plan * e_old[1,:,:]))[1]
-Jx0   = real(1/(Nx*Ny) * (plan * Jx[1,:,:]))[1]
-Jy0   = real(1/(Nx*Ny) * (plan * Jy[1,:,:]))[1]
+#Jx0   = real(1/(Nx*Ny) * (plan * Jx[1,:,:]))[1]
+#Jy0   = real(1/(Nx*Ny) * (plan * Jy[1,:,:]))[1]
 
 #=
 plan = plan_fft(e_old[end,:,:]);
@@ -143,27 +143,34 @@ println("Minimum energy = $(minimum(e[end,:,:]))")
 #println("Old Average Energy Density = $e0_old")
 #println("Old Maximum energy = $(maximum(e_old[end,:,:]))")
 #println("Old Minimum energy = $(minimum(e_old[end,:,:]))")
-#println("A energy = $(e_A[end,1,1])")
-#println("B energy = $(e_B[end,1,1])")
+println("A energy = $(e_A[end,1,1])")
+println("B energy = $(e_B[end,1,1])")
 println("Maximum px = $(maximum(px[end,:,:]))")
 println("Minimum px = $(minimum(px[end,:,:]))")
-println("Average x momenta = $Jx0")
-println("Average y momenta = $Jy0")
-println("Maximum x momenta = $(maximum(abs.(Jx[end,:,:])))")
-println("Maximum y momenta = $(maximum(abs.(Jy[end,:,:])))")
+#println("Average x momenta = $Jx0")
+#println("Average y momenta = $Jy0")
+#println("Maximum x momenta = $(maximum(abs.(Jx[end,:,:])))")
+#println("Maximum y momenta = $(maximum(abs.(Jy[end,:,:])))")
 
-plot(x, e[end,:,25], lw=3)
-plot!(x, e_old[end,15:end-16,40], lw=3)
-xlabel!("x")
 #=
-plot(x,y,e[1,:,:],st=:surface, camera=(50,65))
+a4_PS = BoundaryTimeSeries(PS_dir, :a4)
+a4_A  = BoundaryTimeSeries(A_dir, :a4)[end,1,1]
+a4_B  = BoundaryTimeSeries(B_dir, :a4)[end,1,1]
+a4    = BoundaryTimeSeries(outdir, :a4)
+
+_, xPS, yPS = get_coords(a4_PS, 1, :, :)
+
+
+a4_BB = zeros(length(xPS))
+a4_AA = zeros(length(xPS))
+
+fill!(a4_BB, a4_B)
+fill!(a4_AA, a4_A)
+
+plot(xPS, a4[end, :, 1], lw=3)
+plot!(xPS, a4_PS[end, :, 1], lw=3)
+plot!(xPS, a4_AA, lw=3)
+plot!(xPS, a4_BB, lw=3)
 xlabel!("x")
-ylabel!("y")
+ylabel!("Energy")
 =#
-
-a4     = XiTimeSeries(outdir)
-a4_old = XiTimeSeries(dirname)
-
-plot(x, a4[end,:,25], lw=3)
-plot!(x, a4_old[end,15:end-16,40], lw=3)
-xlabel!("x")
