+++ src/AdS5_3_1/AdS5_3_1.jl
@@ -54,7 +54,8 @@ include("VEVs.jl")
 
 # random utilities useful for post-processing
 include("utils.jl")
-include("design_new_data.jl")
+include("design_new_data_1.jl")
+include("design_new_data_2.jl")
 include("recipes.jl")
 include("GW.jl")
